Move from patch to view override
