+++ castle/cms/browser/content/configure.zcml
@@ -228,4 +228,12 @@
     template="templates/parallaxview.pt"
   />
 
+  <browser:page
+    for="*"
+    name="sharing"
+    class=".actions.SharingView"
+    permission="cmf.ModifyPortalContent"
+    layer="castle.cms.interfaces.ICastleLayer"
+  />
+
 </configure>
