+++ castle/cms/tasks/__init__.py
@@ -4,4 +4,3 @@ from .files import *  # noqa
 from .content import *  # noqa
 from .email import *  # noqa
 from .archive import *  # noqa
-from .sharing import *  # noqa
